pyproject.toml
[project]
name = "ldm-pre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "depth-anything-v2",
    "diffusers[torch]",
    "fsspec>=2026.2.0",
    "hydra-core>=1.3.2",
    "mosaicml-streaming>=0.13.0",
    "pillow>=12.1.1",
    "ray[data]>=2.53.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers<5",
]

[project.scripts]
ldm-pre = "ldm_pre.main:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "third_party/depth-anything-v2",
]

[tool.uv.sources]
depth-anything-v2 = { workspace = true, editable = true }
diffusers = { git = "https://github.com/huggingface/diffusers.git", branch = "main" }
